refactor: move configs to utils/consts
Move all config files and their content to "utils/consts.ts".
