--- src/configs/links.ts
@@ -1,2 +1,0 @@
-export const GITHUB_REPO_LINK =
-    "https://github.com/George-Al3xander/commit-formatter";
