+++ src/utils/helpers/formatString.ts
